Remove select encounter type section from encounter form into its own component
